# PERF: ExtensionBlock.fillna dispatch to underlying arrays